Add scripts for adding Molecules and link them to MoleculeSets

+++ scripts/delete_db_content.py
@@ -1,9 +1,18 @@
 from compounds.models import Molecule, MoleculeSet
 
 
-def run():
-    molecules = Molecule.objects.all()
-    molecules.delete()
+def run(*args) -> None:
+    """
+    Deletes all Molecule and MoleculeSet instances in the DB
+    :param args: list of arguments
+                 no-sets: exclude all MoleculeSets
+                 no-molecules: exclude all Molecules
+    :return: None
+    """
+    if 'no-molecules' not in args:
+        molecules = Molecule.objects.all()
+        molecules.delete()
 
-    molecule_sets = MoleculeSet.objects.all()
-    molecule_sets.delete()
+    if 'no-sets' not in args:
+        molecule_sets = MoleculeSet.objects.all()
+        molecule_sets.delete()
